pyproject.toml
[project]
name = "python-crop-rects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.1.0",
    "pymupdf>=1.26.7",
]
